+++ tf/enterpriseproj.tf
@@ -10,3 +10,7 @@ data "huaweicloud_enterprise_project" "ep_prod" {
   name = "PROD"
 }
 
+data "huaweicloud_enterprise_project" "ep_net" {
+  name = "NET"
+}
+