+++ tf/vpc_subnet.tf
@@ -22,6 +22,14 @@ resource "huaweicloud_vpc" "prod" {
   enterprise_project_id = data.huaweicloud_enterprise_project.ep_prod.id
 }
 
+resource "huaweicloud_vpc" "net" {
+  region  = "sa-brazil-1"
+  name    = "NET"
+  cidr    = "10.4.0.0/16"
+  description = "Inter-network"
+  enterprise_project_id = data.huaweicloud_enterprise_project.ep_prod.id
+}
+
 resource "huaweicloud_vpc_subnet" "dev_dev01" {
   region      = huaweicloud_vpc.dev.region
   name        = "DEV01"
@@ -48,6 +56,16 @@ resource "huaweicloud_vpc_subnet" "prod_prod01" {
   cidr        = "10.3.1.0/24"
   gateway_ip  = "10.3.1.1"
   vpc_id      = huaweicloud_vpc.prod.id
+  description = ""
+  availability_zone = "sa-brazil-1a"
+}
+
+resource "huaweicloud_vpc_subnet" "net_net01" {
+  region      = huaweicloud_vpc.net.region
+  name        = "NET01"
+  cidr        = "10.4.1.0/24"
+  gateway_ip  = "10.4.1.1"
+  vpc_id      = huaweicloud_vpc.net.id
   description = ""
   availability_zone = "sa-brazil-1a"
 }