+++ tf/ecs.tf
@@ -94,6 +94,10 @@ resource "huaweicloud_compute_instance" "ecs_stg" {
     fixed_ip_v4       = "10.2.1.21"
     source_dest_check = true
   }
+
+  scheduler_hints {
+    tenancy       = "dedicated"
+  }
 }
 
 resource "huaweicloud_compute_servergroup" "servergroup_group01" {