+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
-- **LLD**: Add Enterprise Project column to security group;
+- **LLD**: Add Enterprise Project column to Security Group tag;
 - **LLD**: Renamed "Enterprise Project Name" column in ECS tab to "Enterprise
   Project";
+- **LLD**: Removed "Password or Keypair" column in ECS tab, because "OS
+  Password" is mandatory in order to set a keypair;
 - Add Enterprise Project support to VPC, security group, ECS and EVS;
 - Refactored servergroup dependencies, now creating servergroup first and then
   configuring their ID in ECS scheduler hints. This also removed the dependency