+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - **LLD**: Removed "Password or Keypair" column in ECS tab, because "OS
   Password" is mandatory in order to set a keypair;
 - **metadata**: Added "ECS last wave" parameter;
-- **LLD**: ECS Wave column now mandatory, if wave is negative or larger than
-  "ECS last wave" parameter, the ECS will be ignored;
+- **LLD**: "Wave" column in ECS tab is now mandatory; if wave is negative or
+  larger than "ECS last wave" parameter, the ECS will be ignored;
+- **LLD**: Add "Dedicated Host" column to ECS tab;
 - Add Enterprise Project support to VPC, security group, ECS and EVS;
 - Refactored servergroup dependencies, now creating servergroup first and then
   configuring their ID in ECS scheduler hints. This also removed the dependency